Refactor of V3 controller methods
