+++ plugin.php
@@ -1,14 +1,12 @@
 <?php
 /*
 Plugin Name: Ipswich JAFFA RC Results WP REST API
-* Ref: https://github.com/petenelson/extending-wp-rest-api
 */
 
 if ( ! defined( 'ABSPATH' ) ) die( 'restricted access' );
 
 require_once plugin_dir_path( __FILE__ ) .'WordPressApiHelper.php';
 
-require_once plugin_dir_path( __FILE__ ) .'v2/AccountController.php';
 require_once plugin_dir_path( __FILE__ ) .'v2/EventsController.php';
 require_once plugin_dir_path( __FILE__ ) .'v2/DistancesController.php';
 require_once plugin_dir_path( __FILE__ ) .'v2/MeetingsController.php';
@@ -19,12 +17,13 @@ require_once plugin_dir_path( __FILE__ ) .'v2/RunnersController.php';
 require_once plugin_dir_path( __FILE__ ) .'v2/StatisticsController.php';
 require_once plugin_dir_path( __FILE__ ) .'v2/LeaguesController.php';
 
-require_once plugin_dir_path( __FILE__ ) .'v3/class-ipswich-jaffa-results-wp-rest-api-controller-v3.php';
+require_once plugin_dir_path( __FILE__ ) .'v3/ResultsController.php';
+require_once plugin_dir_path( __FILE__ ) .'v3/RunnerOfTheMonthController.php';
+require_once plugin_dir_path( __FILE__ ) .'v3/AdminController.php';
 
 // hook into the rest_api_init action so we can start registering routes
 $namespace = 'ipswich-jaffa-api/v2'; // base endpoint for our custom API
 
-$accountController = new IpswichJAFFARunningClubAPI\V2\AccountController($namespace);
 $eventsController = new IpswichJAFFARunningClubAPI\V2\EventsController($namespace);
 $distancesController = new IpswichJAFFARunningClubAPI\V2\DistancesController($namespace);
 $meetingsController = new IpswichJAFFARunningClubAPI\V2\MeetingsController($namespace);
@@ -34,6 +33,12 @@ $runnerOfTheMonthController = new IpswichJAFFARunningClubAPI\V2\RunnerOfTheMonth
 $runnersController = new IpswichJAFFARunningClubAPI\V2\RunnersController($namespace);
 $statisticsController = new IpswichJAFFARunningClubAPI\V2\StatisticsController($namespace);
 $leaguesController = new IpswichJAFFARunningClubAPI\V2\LeaguesController($namespace);
+
+$namespaceV3 = 'ipswich-jaffa-api/v3'; 
+$adminV3Controller = new IpswichJAFFARunningClubAPI\V3\AdminController($namespaceV3);
+$resultsV3Controller = new IpswichJAFFARunningClubAPI\V3\ResultsController($namespaceV3);
+$runnerOfTheMonthV3Controller = new IpswichJAFFARunningClubAPI\V3\RunnerOfTheMonthController($namespaceV3);
+
 $helper = new IpswichJAFFARunningClubAPI\WordPressApiHelper();
 
 add_action( 'rest_api_init', array( $accountController, 'registerRoutes') );
@@ -47,11 +52,10 @@ add_action( 'rest_api_init', array( $runnersController, 'registerRoutes') );
 add_action( 'rest_api_init', array( $statisticsController, 'registerRoutes') );
 add_action( 'rest_api_init', array( $leaguesController, 'registerRoutes') );
 
-$api_controller_V3 = new IpswichJAFFARunningClubAPI\V3\Ipswich_JAFFA_Results_WP_REST_API_Controller_V3();
-add_action( 'rest_api_init', array( $api_controller_V3, 'rest_api_init') );
+add_action( 'rest_api_init', array( $resultsV3Controller, 'registerRoutes') );
+add_action( 'rest_api_init', array( $adminV3Controller, 'registerRoutes') );
+add_action( 'rest_api_init', array( $runnerOfTheMonthV3Controller, 'registerRoutes') );
 
 // Customise user response for JWT login
 add_filter( 'jwt_auth_token_before_dispatch', array($helper, 'custom_wp_user_token_response'), 10, 2);
-
-add_action( 'plugins_loaded', array( $helper, 'pluginsLoaded') );
 ?>
