+++ v3/IRoute.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace IpswichJAFFARunningClubAPI\V3;
+
+interface IRoute {
+  public function registerRoutes();
+}
+?>
